backend/requirements.txt
# Core
fastapi
uvicorn[standard]
python-dotenv

# RAG
langchain
langchain-community
langchain-groq
langchain-text-splitters
sentence-transformers
chromadb-client
pypdf

# Agent
langgraph

# Optional for deployment
gunicorn
